Move lots of stuff to atlasdb-cassandra
