+++ atlasdb-backup/build.gradle
@@ -15,6 +15,4 @@ dependencies {
 
     annotationProcessor group: 'org.immutables', name: 'value'
     compileOnly 'org.immutables:value::annotations'
-    annotationProcessor 'org.derive4j:derive4j'
-    compileOnly 'org.derive4j:derive4j-annotation'
 }
