+++ atlasdb-backup/src/main/java/com/palantir/atlasdb/backup/AtlasRestoreService.java
@@ -16,30 +16,20 @@
 
 package com.palantir.atlasdb.backup;
 
-import com.datastax.driver.core.Metadata;
-import com.datastax.driver.core.Session;
-import com.datastax.driver.core.Token;
-import com.datastax.driver.core.TokenRange;
-import com.datastax.driver.core.policies.DefaultRetryPolicy;
 import com.google.common.annotations.VisibleForTesting;
 import com.google.common.collect.Range;
 import com.palantir.atlasdb.backup.api.CompletedBackup;
-import com.palantir.atlasdb.backup.transaction.TransactionsTableInteraction;
 import com.palantir.atlasdb.cassandra.CassandraKeyValueServiceConfig;
 import com.palantir.atlasdb.cassandra.backup.CassandraRepairHelper;
-import com.palantir.atlasdb.cassandra.backup.ClusterMetadataUtils;
 import com.palantir.atlasdb.internalschema.InternalSchemaMetadataState;
 import com.palantir.atlasdb.keyvalue.api.KeyValueService;
 import com.palantir.atlasdb.keyvalue.cassandra.LightweightOppToken;
 import com.palantir.atlasdb.timelock.api.Namespace;
-import com.palantir.common.streams.KeyedStream;
 import com.palantir.logsafe.SafeArg;
 import com.palantir.logsafe.logger.SafeLogger;
 import com.palantir.logsafe.logger.SafeLoggerFactory;
 import com.palantir.timestamp.FullyBoundedTimestampRange;
 import java.net.InetSocketAddress;
-import java.util.HashSet;
-import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 import java.util.Set;
@@ -110,69 +100,7 @@ public class AtlasRestoreService {
         Map<FullyBoundedTimestampRange, Integer> coordinationMap =
                 CoordinationServiceUtilities.getCoordinationMapOnRestore(
                         schemaMetadataState, fastForwardTs, immutableTs);
-        List<TransactionsTableInteraction> transactionsTableInteractions =
-                TransactionsTableInteraction.getTransactionTableInteractions(
-                        coordinationMap, DefaultRetryPolicy.INSTANCE);
-
-        // 3. get partition tokens
-        Map<String, Set<Token>> partitionKeysByTable =
-                getPartitionKeysByTable(namespace, transactionsTableInteractions);
-
-        if (log.isDebugEnabled()) {
-            Map<String, FullyBoundedTimestampRange> loggableTableRanges = KeyedStream.of(transactionsTableInteractions)
-                    .mapKeys(TransactionsTableInteraction::getTransactionsTableName)
-                    .map(TransactionsTableInteraction::getTimestampRange)
-                    .collectToMap();
-            Map<String, Integer> numPartitionKeysByTable =
-                    KeyedStream.stream(partitionKeysByTable).map(Set::size).collectToMap();
-            log.debug(
-                    "Identified token ranges requiring repair in the following transactions tables",
-                    SafeArg.of("transactionsTablesWithRanges", loggableTableRanges),
-                    SafeArg.of("numPartitionKeysByTable", numPartitionKeysByTable));
-        }
-
-        // 4. get repair ranges
-        // TODO(gs): don't recreate a session here, we already got this information before
-        Map<String, Map<InetSocketAddress, Set<Range<LightweightOppToken>>>> tokenRangesForRepair;
-        try (Session session = cassandraRepairHelper.newSession(namespace)) {
-            String keyspaceName = session.getLoggedKeyspace();
-            Metadata metadata = session.getCluster().getMetadata();
-
-            Set<InetSocketAddress> hosts = new HashSet<>(); // TODO(gs): move to somewhere where we have the hosts!!
-            tokenRangesForRepair = KeyedStream.stream(partitionKeysByTable)
-                    .map(ranges -> ClusterMetadataUtils.getTokenMapping(hosts, metadata, keyspaceName, ranges))
-                    .map(this::makeLightweight)
-                    .collectToMap();
-        }
-
-        // 5. repair ranges
-        KeyedStream.stream(tokenRangesForRepair).forEach((table, ranges) -> {
-            log.info("Repairing ranges for table", SafeArg.of("table", table));
-            repairTable.accept(ranges);
-        });
-    }
-
-    // TODO(gs): ugh. this gives us a stream in a stream in a stream
-    private Map<InetSocketAddress, Set<Range<LightweightOppToken>>> makeLightweight(
-            Map<InetSocketAddress, Set<TokenRange>> ranges) {
-        return KeyedStream.stream(ranges)
-                .map(cassandraRepairHelper::makeLightweight)
-                .collectToMap();
-    }
-
-    private Map<String, Set<Token>> getPartitionKeysByTable(
-            Namespace namespace, List<TransactionsTableInteraction> transactionsTableInteractions) {
-        try (Session session = cassandraRepairHelper.newSession(namespace)) {
-            String keyspaceName = session.getLoggedKeyspace();
-            Metadata metadata = session.getCluster().getMetadata();
-            return KeyedStream.of(transactionsTableInteractions.stream())
-                    .mapKeys(TransactionsTableInteraction::getTransactionsTableName)
-                    .map(interaction -> interaction.getPartitionTokens(
-                            ClusterMetadataUtils.getTableMetadata(
-                                    metadata, keyspaceName, interaction.getTransactionsTableName()),
-                            session))
-                    .collectToMap();
-        }
+        cassandraRepairHelper.repairTransactionsTables(namespace, coordinationMap, repairTable);
     }
 
     private boolean backupExists(Namespace namespace) {
