+++ atlasdb-cassandra/src/main/java/com/palantir/atlasdb/cassandra/backup/transaction/Transactions1TableInteraction.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.palantir.atlasdb.backup.transaction;
+package com.palantir.atlasdb.cassandra.backup.transaction;
 
 import com.datastax.driver.core.BoundStatement;
 import com.datastax.driver.core.ConsistencyLevel;
@@ -27,26 +27,28 @@ import com.datastax.driver.core.policies.DefaultRetryPolicy;
 import com.datastax.driver.core.policies.RetryPolicy;
 import com.datastax.driver.core.querybuilder.QueryBuilder;
 import com.datastax.driver.core.utils.Bytes;
+import com.google.common.collect.ImmutableList;
 import com.palantir.atlasdb.keyvalue.api.Cell;
 import com.palantir.atlasdb.keyvalue.cassandra.CassandraConstants;
-import com.palantir.atlasdb.pue.KvsConsensusForgettingStore;
-import com.palantir.atlasdb.pue.PutUnlessExistsValue;
-import com.palantir.atlasdb.transaction.encoding.TwoPhaseEncodingStrategy;
+import com.palantir.atlasdb.transaction.encoding.V1EncodingStrategy;
 import com.palantir.atlasdb.transaction.impl.TransactionConstants;
 import com.palantir.timestamp.FullyBoundedTimestampRange;
 import java.nio.ByteBuffer;
 import java.util.List;
-import java.util.Set;
-import java.util.stream.Collectors;
 
-public class Transactions3TableInteraction implements TransactionsTableInteraction {
+public class Transactions1TableInteraction implements TransactionsTableInteraction {
     private final FullyBoundedTimestampRange timestampRange;
     private final RetryPolicy abortRetryPolicy;
 
-    public Transactions3TableInteraction(FullyBoundedTimestampRange timestampRange, RetryPolicy abortRetryPolicy) {
+    public Transactions1TableInteraction(FullyBoundedTimestampRange timestampRange, RetryPolicy abortRetryPolicy) {
         this.timestampRange = timestampRange;
         this.abortRetryPolicy = abortRetryPolicy;
     }
+
+    private static final byte[] DUMMY = new byte[] {1};
+    private static final byte[] ABORT_COMMIT_TS_ENCODED =
+            V1EncodingStrategy.INSTANCE.encodeCommitTimestampAsValue(0, TransactionConstants.FAILED_COMMIT_TS);
+    static final ByteBuffer COLUMN_NAME_BB = ByteBuffer.wrap(TransactionConstants.COMMIT_TS_COLUMN);
 
     @Override
     public FullyBoundedTimestampRange getTimestampRange() {
@@ -55,18 +57,15 @@ public class Transactions3TableInteraction implements TransactionsTableInteracti
 
     @Override
     public String getTransactionsTableName() {
-        return TransactionConstants.TRANSACTIONS2_TABLE.getTableName();
+        return TransactionConstants.TRANSACTION_TABLE.getTableName();
     }
 
     @Override
     public PreparedStatement prepareAbortStatement(TableMetadata transactionsTable, Session session) {
-        // we are declaring bankruptcy if this fails anyway
-        ByteBuffer abortCommitTsBb = ByteBuffer.wrap(TwoPhaseEncodingStrategy.ABORTED_TRANSACTION_COMMITTED_VALUE);
-
         Statement abortStatement = QueryBuilder.update(transactionsTable)
-                .with(QueryBuilder.set(CassandraConstants.VALUE, abortCommitTsBb))
+                .with(QueryBuilder.set(CassandraConstants.VALUE, ByteBuffer.wrap(ABORT_COMMIT_TS_ENCODED)))
                 .where(QueryBuilder.eq(CassandraConstants.ROW, QueryBuilder.bindMarker()))
-                .and(QueryBuilder.eq(CassandraConstants.COLUMN, QueryBuilder.bindMarker()))
+                .and(QueryBuilder.eq(CassandraConstants.COLUMN, COLUMN_NAME_BB))
                 .and(QueryBuilder.eq(CassandraConstants.TIMESTAMP, CassandraConstants.ENCODED_CAS_TABLE_TIMESTAMP))
                 .onlyIf(QueryBuilder.eq(CassandraConstants.VALUE, QueryBuilder.bindMarker()));
         // if you change this from CAS then you must update RetryPolicy
@@ -78,32 +77,25 @@ public class Transactions3TableInteraction implements TransactionsTableInteracti
         Statement checkStatement = QueryBuilder.select()
                 .from(transactionsTable)
                 .where(QueryBuilder.eq(CassandraConstants.ROW, QueryBuilder.bindMarker()))
-                .and(QueryBuilder.eq(CassandraConstants.COLUMN, QueryBuilder.bindMarker()))
+                .and(QueryBuilder.eq(CassandraConstants.COLUMN, COLUMN_NAME_BB))
                 .and(QueryBuilder.eq(CassandraConstants.TIMESTAMP, CassandraConstants.ENCODED_CAS_TABLE_TIMESTAMP));
         return session.prepare(checkStatement.toString());
     }
 
     @Override
     public TransactionTableEntry extractTimestamps(Row row) {
-        long startTimestamp = TwoPhaseEncodingStrategy.INSTANCE.decodeCellAsStartTimestamp(Cell.create(
-                Bytes.getArray(row.getBytes(CassandraConstants.ROW)),
-                Bytes.getArray(row.getBytes(CassandraConstants.COLUMN))));
-        PutUnlessExistsValue<Long> commitValue = TwoPhaseEncodingStrategy.INSTANCE.decodeValueAsCommitTimestamp(
-                startTimestamp, Bytes.getArray(row.getBytes(CassandraConstants.VALUE)));
-        if (commitValue.value() == TransactionConstants.FAILED_COMMIT_TS) {
-            return TransactionTableEntries.explicitlyAborted(startTimestamp);
-        }
-
-        return TransactionTableEntries.committedTwoPhase(startTimestamp, commitValue);
+        long startTimestamp = decodeStartTs(Bytes.getArray(row.getBytes(CassandraConstants.ROW)));
+        long commitTimestamp = decodeCommitTs(Bytes.getArray(row.getBytes(CassandraConstants.VALUE)));
+        return commitTimestamp == TransactionConstants.FAILED_COMMIT_TS
+                ? TransactionTableEntries.explicitlyAborted(startTimestamp)
+                : TransactionTableEntries.committedLegacy(startTimestamp, commitTimestamp);
     }
 
     @Override
     public Statement bindCheckStatement(PreparedStatement preparedCheckStatement, TransactionTableEntry entry) {
         long startTs = TransactionTableEntries.getStartTimestamp(entry);
-        Cell cell = TwoPhaseEncodingStrategy.INSTANCE.encodeStartTimestampAsCell(startTs);
-        ByteBuffer rowKeyBb = ByteBuffer.wrap(cell.getRowName());
-        ByteBuffer columnNameBb = ByteBuffer.wrap(cell.getColumnName());
-        BoundStatement bound = preparedCheckStatement.bind(rowKeyBb, columnNameBb);
+        ByteBuffer startTimestampBb = encodeStartTimestamp(startTs);
+        BoundStatement bound = preparedCheckStatement.bind(startTimestampBb);
         return bound.setConsistencyLevel(ConsistencyLevel.QUORUM)
                 .setSerialConsistencyLevel(ConsistencyLevel.SERIAL)
                 .setReadTimeoutMillis(LONG_READ_TIMEOUT_MS)
@@ -113,17 +105,13 @@ public class Transactions3TableInteraction implements TransactionsTableInteracti
     @Override
     public Statement bindAbortStatement(PreparedStatement preparedAbortStatement, TransactionTableEntry entry) {
         long startTs = TransactionTableEntries.getStartTimestamp(entry);
-        PutUnlessExistsValue<Long> commitTs =
-                TransactionTableEntries.getCommitValue(entry).orElseThrow(() -> illegalEntry(entry));
-        Cell cell = TwoPhaseEncodingStrategy.INSTANCE.encodeStartTimestampAsCell(startTs);
-        ByteBuffer rowKeyBb = ByteBuffer.wrap(cell.getRowName());
-        ByteBuffer columnNameBb = ByteBuffer.wrap(cell.getColumnName());
-        ByteBuffer valueBb =
-                ByteBuffer.wrap(TwoPhaseEncodingStrategy.INSTANCE.encodeCommitTimestampAsValue(startTs, commitTs));
-        BoundStatement bound = preparedAbortStatement.bind(rowKeyBb, columnNameBb, valueBb);
+        long commitTs = TransactionTableEntries.getCommitTimestamp(entry).orElseThrow(() -> illegalEntry(entry));
+        ByteBuffer startTimestampBb = encodeStartTimestamp(startTs);
+        ByteBuffer commitTimestampBb = encodeCommitTimestamp(commitTs);
+        BoundStatement bound = preparedAbortStatement.bind(startTimestampBb, commitTimestampBb);
         return bound.setConsistencyLevel(ConsistencyLevel.QUORUM)
                 .setSerialConsistencyLevel(ConsistencyLevel.SERIAL)
-                .setDefaultTimestamp(KvsConsensusForgettingStore.PUT_TIMESTAMP + 1)
+                .setDefaultTimestamp(CassandraConstants.CAS_TABLE_TIMESTAMP)
                 .setReadTimeoutMillis(LONG_READ_TIMEOUT_MS)
                 .setIdempotent(true) // by default CAS operations are not idempotent in case of multiple clients
                 .setRetryPolicy(abortRetryPolicy);
@@ -131,19 +119,37 @@ public class Transactions3TableInteraction implements TransactionsTableInteracti
 
     @Override
     public List<Statement> createSelectStatements(TableMetadata transactionsTable) {
-        Set<ByteBuffer> encodedRowKeys = TwoPhaseEncodingStrategy.INSTANCE
-                .encodeRangeOfStartTimestampsAsRows(
-                        timestampRange.inclusiveLowerBound(), timestampRange.inclusiveUpperBound())
-                .map(ByteBuffer::wrap)
-                .collect(Collectors.toSet());
-        return encodedRowKeys.stream()
-                .map(rowKey -> QueryBuilder.select()
-                        .all()
-                        .from(transactionsTable)
-                        .where(QueryBuilder.eq(CassandraConstants.ROW, rowKey))
-                        .setConsistencyLevel(ConsistencyLevel.QUORUM)
-                        .setFetchSize(SELECT_TRANSACTIONS_FETCH_SIZE)
-                        .setReadTimeoutMillis(LONG_READ_TIMEOUT_MS))
-                .collect(Collectors.toList());
+        Statement select = QueryBuilder.select()
+                .all()
+                .from(transactionsTable)
+                .where(QueryBuilder.lte(
+                        QueryBuilder.token(CassandraConstants.ROW),
+                        QueryBuilder.token(encodeStartTimestamp(timestampRange.inclusiveUpperBound()))))
+                .and(QueryBuilder.gte(
+                        QueryBuilder.token(CassandraConstants.ROW),
+                        QueryBuilder.token(encodeStartTimestamp(timestampRange.inclusiveLowerBound()))))
+                .setConsistencyLevel(ConsistencyLevel.QUORUM)
+                .setFetchSize(SELECT_TRANSACTIONS_FETCH_SIZE)
+                .setReadTimeoutMillis(LONG_READ_TIMEOUT_MS);
+
+        return ImmutableList.of(select);
+    }
+
+    static ByteBuffer encodeStartTimestamp(long timestamp) {
+        return ByteBuffer.wrap(V1EncodingStrategy.INSTANCE
+                .encodeStartTimestampAsCell(timestamp)
+                .getRowName());
+    }
+
+    private long decodeStartTs(byte[] startTsRow) {
+        return V1EncodingStrategy.INSTANCE.decodeCellAsStartTimestamp(Cell.create(startTsRow, DUMMY));
+    }
+
+    static ByteBuffer encodeCommitTimestamp(long timestamp) {
+        return ByteBuffer.wrap(V1EncodingStrategy.INSTANCE.encodeCommitTimestampAsValue(0, timestamp));
+    }
+
+    private long decodeCommitTs(byte[] value) {
+        return V1EncodingStrategy.INSTANCE.decodeValueAsCommitTimestamp(0, value);
     }
 }
