+++ atlasdb-cassandra/src/test/java/com/palantir/atlasdb/cassandra/backup/transaction/TransactionTableEntryAssertions.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.palantir.atlasdb.backup.transaction;
+package com.palantir.atlasdb.cassandra.backup.transaction;
 
 import static org.assertj.core.api.Assertions.fail;
